requirements.txt
asgiref==3.3.4
Django==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
